CREP-12 Move signup code to signup.js
